requirements.txt
streamlit==1.22.0
tensorflow==2.10.0
numpy==1.23.5
Pillow==9.3.0
opencv-python==4.6.0.66  # Only if processing videos